Default to using release build types
This changes the default build type to be "release"
https://mesonbuild.com/Builtin-options.html#details-for-buildtype

This optimizes the build and removes debug symbols.

I don't see a speed difference with my timers, but we probably have more
work to do there. The build is 65% smaller.

I also added docs for how to make a debug build for real debugging.
We can also use the `#if !NDBUG` macro (meson adds that for free?) for
any code we want to gate on a debug build.